feat: 🎸 extract image style
